move getSecurityModules function out of extension
