+++ src/utils.ts
@@ -338,3 +338,40 @@ export function moveTempFolder(previousTempPath: string, currentTempPath: string
         vscode.window.showErrorMessage(err.message);
     }
 }
+
+/**
+ * function to check if imperative.json contains
+ * information about security or not and then
+ * Imports the neccesary security modules
+ */
+export function getSecurityModules(moduleName): NodeRequire | undefined {
+  let imperativeIsSecure: boolean = false;
+  try {
+      const fileName = path.join(getZoweDir(), "settings", "imperative.json");
+      let settings: any;
+      if (fs.existsSync(fileName)) {
+          settings = JSON.parse(fs.readFileSync(fileName).toString());
+      }
+      const value1 = settings?.overrides.CredentialManager;
+      const value2 = settings?.overrides["credential-manager"];
+      imperativeIsSecure = ((typeof value1 === "string") && (value1.length > 0)) ||
+          ((typeof value2 === "string") && (value2.length > 0));
+  } catch (error) {
+      globals.LOG.warn(localize("profile.init.read.imperative", "Unable to read imperative file. ") + error.message);
+      vscode.window.showWarningMessage(error.message);
+      return undefined;
+  }
+  if (imperativeIsSecure) {
+      // Workaround for Theia issue (https://github.com/eclipse-theia/theia/issues/4935)
+      const appRoot = globals.ISTHEIA ? process.cwd() : vscode.env.appRoot;
+      try {
+          return require(`${appRoot}/node_modules/${moduleName}`);
+      } catch (err) { /* Do nothing */ }
+      try {
+          return require(`${appRoot}/node_modules.asar/${moduleName}`);
+      } catch (err) { /* Do nothing */ }
+      vscode.window.showWarningMessage(localize("initialize.module.load",
+          "Credentials not managed, unable to load security file: ") + moduleName);
+  }
+  return undefined;
+}
